out-comment length method for RLum.Results objects
Caused all sorts of errors. In the first place, the problem was that not all RLum.Results objects had an @data$data slot, so the length was NULL (which in turn RStudio does not like).
Passing it just to length(x) caused the error
Error: protect () : protection stack overflow

No idea why.. anyhow, we should come up with a meaningful method first