fix calculation se(OD)
This bug was introduced by RG himself in the original scripts, where se(OD) was calculated as 1/sqrt(2*sig*sum(wi)). Galbraith et al. 1999 and Galbraith & Roberts 2012 clearly state, however, that it should be 1/sqrt(2*sig^2*sum(wi)).
The latter is now implemented and was tested and cross-checked with the worked example given in Galbraith et al. 1999.