fix/improve channel calculation of L3
see commit 69cb9915

Problem: read_BIN2R (probably) imports the OSL curves and sets the stimulation time from 0 to max stim.-time. Hence, the channel width is larger than the difference in time between channels. The last time data point should be time - channel width. For a 2 s SG measurement with 110 channels the first and last time should be 0 and 1.982 (and not 2!).