Experimental: get cross-sections from fit_CWCurve fitting
May be removed in the future; or be enhance by also deriving the fast ratio from the nls-fit returned from fit_CWCurve fitting. The latter may yield more robust results as the curve is smoothed. In low-intensity and and low-resolution curves deriving the counts from just one channel for L2 may significantly alter the results due to counting statistics.