analyse_SAR.CWOSL: DRT compatibility
In dose recovery tests (DRT) the natural signal is often removed not within the device, but under sunlight or in a SOL2. The first OSL/IRSL curve in the BIN is the regenerated signal with a given beta dose that is treated as the natural. The METAFILE stores the previous irradiation for this file, even though we want to treat it as the "natural". We now check whether the first dose is 0 and if not, set it to 0.

(cherry picked from commit 26b3e2e5b8f5f1370350527cfeda751319ab5147)