use_DRAC: assign correct class to output elements #27
More of a makeshift fix, but works reliably. Before, read.table was not able to automatically detect the column class because the first two lines (header) were included. Now, we read the raw string in twice, the second time skipping the header lines so read.table is able to determine the correct class.