remove DOT from Rcpp exported functions
@RLumSK
There are/were some weird issues, ultimately making it impossible to compile and build the package on Windows. Given previous experiences and considering that a DOT in function names has a different semantic meaning in C++ and R we are probably best of removing this stylistic choice. Compiling and building the package as of this commit works fine on Windows, but please verify for Mac.