misc: fix 'missing links' when building html help.
During installation of the package the help indices are converted to html, which produced a couple of "missing link" errors. This is now fixed by proper referencing of the help files using the double square brackets syntax of the roxygen2 markdown feature.

NB: these changes require a full rebuild. updated Rd files not included for clarity.