rlum.analysis: remove check for valid subset expression
The check produces a non-meaningful and confusing error if the subset variable is not found within the environment. Any error produced is caught anyway by the subsequent call to tryCatch(eval()) anyway, so I do not see a good reason why we should keep this check.

If the user provides a non-logical expression for "subset" in subset(), argument "subset" in get_RLum() is NULL and no subsetting occurs. So the additional benefit of a more meaningful error is only when get_RLum(<obj>, subset = (<condition>)) is called. But this would require additional checks on what the calling frame is.