drac: do not treat "F" for mineral as logical (fixes #68)
read.table() treats the character string "F" as the global variable for FALSE. To avoid this behavior we need to enforce column classes. By using data.table::fread() instead we avoid the need to specify the class for each column, as it allows enforcing the class only for certain columns.