plotDRC: allow analyse_pIRIR data and add return value
RLum.Results objects produced by analyse_pIRIRSequence() work just as well for plot_DRCSummary(), so we can allow its originator here.
Also added a return object (RLum.Results) that mainly contains a data.frame with the dose in the first column and the lxtx values of the curves in all subsequent columns.