Fix computation of skewness and kurtosis with MCM.
When `n.MCM` is not NULL, `data.MCM` is a matrix with as many
columns as number of Monte Carlo samples: calling `sum()` over
it when computating skewness and kurtosis will sum `n * n.MCM`
values, so we must divide by `n * n.MCM` to produce the correct
expectations.

Fixes #122.