Copy the tests/testdata files to tests/testthat/_data and package them.
The new directory is now being included into the package (with exclusion
of SPEfile.SPE as much larger than the rest). File BINfile.V8.binx is not
copied as an identical file is already under inst/extdata and that's what
is already being used in the read_BIN2R() tests.

This is a preliminary step before the tests/testdata directory can be
removed and the test files updated.