Consolidate code in plot_DetPlot().
Before #206, signal.integral.min == signal.integral.max was allowed, so
this case had to be dealt with in a special way when constructing the
signal_integral.seq vector: in particular, it could not use directly

   seq(..., by = signal.integral.max - signal.integral.min)

as that evaluated to seq(..., by = 0). So it constructed manually a range,
which effectively corresponds to using seq(..., by = 1).

Given that signal.integral.min == signal.integral.max is not allowed
anymore, we can consolidate the code leaving only the seq() version.