Rework the creation of RejectionCriteria in analyse_SAR.TL().
The code that creates the RejectionCriteria data frame has been entirely
rewritten using data.table, resulting in cleaner and also more performant
code.

Moreover, this change allows to solve an additional testcase:

  analyse_SAR.TL(object, dose.points = 2,
                 signal.integral.min = 210, signal.integral.max = 220,
                 sequence.structure = c("SIGNAL", "BACKGROUND"))

which previously failed with:

  Error in dimnames(x) <- dn :
    length of 'dimnames' [2] not equal to array extent

due to the use of sapply(), which in this case ended up producing a matrix
instead of a vector.