Don't crash in calc_AverageDose() on some RLum.Results objects.
The RLum.Results object generated by calc_AverageDose() has only one row,
and R returns its standard deviation as NA. In this case, we reset it to 0
and carry on. However, with just one value, also the breaks argument of
hist() needs to be manually set to avoid an error.