Adapt the plot_GrowthCurve() snapshot tests to run on R 4.3.3.
The HDPI_L and HPDI_U values are calculated in .calc_HPDI(), which uses
density(): as there has been a change in its implementation in R 4.4.0,
the values produced are different.

While we use R 4.3.3 in CI, we remove those values from the snapshot, so
we can run those tests on all Linux versions as well as getting coverage
from them.