Return also the computed De without abs() from fit_DoseResponseCurve().
This allows to produce correct plots in the extrapolation case, where De
can be negative.

This is now required because plotting happens after fitting, while in the
old plot_GrowthCurve() prior to 089c136 plotting occurred before it, so the
computed De (which could be negative) could be used directly in the plot.