Streamline the reading of strings in read_BIN2R().
This adds an inner function to consolidates the code to read the strings.

Each string is now read up to its total allocated size with one single
call to readBin(), while the call to readChar() is done on the raw vector
up to the used size (or forced size for some time/date fields): this
avoids calls to seek.connection() that were before necessary when we were
only reading the used part of the field.