Avoid warnings when max() returns -Inf.
It may happen that we try to calculate the max() of a vector that only
contains NAs: if so, we reset Max to 1 so that the algorithm will throw
the numerical error warning and stop soon after.

This is just a bandaid fix, presumably there is a better way to solve
this.