Speed up calc_Huntley2006(): set a more consistent default for rprime.
Before the number of points in rprime was 500 for the natural dose while
it depended on the number of MC points for the computation of nN_SS;
moreover its upper boundary was at 5, but wetting it to 3 allows to
concentrate more points in the bulk of the distribution. The new default
is to have 1000 equally-spaced points between 0.01 and 3, but this is
configurable via the 'rprime' argument.

This brings a very large speedup, especially noticeable for large
values of n_MC, with only very minor differences in the results.
In the testcase with n_MC = 50000, the time went from 137s to 60s.