Fix placement of the summary text in the DRC plot from calc_Huntley().
Previouly the placement was based on the last tick on the x-axis, but that
could lead to the summary text to overlap the main plot. Now we place it
according to the rightmost point in the plot, so overlaps should not occur
anymore.

Thanks to @SalOehl for reporting the issue and providing a test dataset.