Fix regression in plot generated by analyse_FadingMeasurement().
The issue seems to be that setting mfrow resets the par settings (as the
pre-existing comment stated) even in the case when mfrow is not actually
changed. Therefore, we cannot have a shared par.default for the two cases
of plot_singlePanels, but each must define its own.

This fixes a regression introduced in 9f5a918.