Verify that the number of unique sample names matches what we expect.
The number of unique sample names must match the number of IRSL/OSL ratios
we compute, as otherwise it would indicate some input misspecification.

This avoids the following error:

 Error in data.frame(ID = seq_along(OSL$sum_signal), RUN = run, BSL = OSL$sum_signal, :
   arguments imply differing number of rows: 13, 1, 14

which could happen if a sample was given different coordinates in its sample
name.

This check revealed that the way we used to define the coordinates for the
surface object in the tests was incorrect, as it attached multiple sets of
coordinates to the same sample, thus triggering the new check.