Make fit_ThermalQuenching() more stable.
This adopts a slight reformulation that is mathematically equivalent to
the one used before. As it operates on the log-space for the C parameter,
it helps to set an upper bound that's large enough but not infinite, to
avoid possible overflows when exponentiating the parameter to return it
to the original scale.