Make the estimation of the y-scaling in calc_FiniteMixture() more robust.
Rather than doing

  na.exclude(comp.n[pos.n + 1, ])

we now do

  na.exclude(c(comp.n[pos.n + 1, ]))

so that we only remove the NA elements from a vector, rather than removing
any row that contains NAs, as that may leave a vector of length 0.

This code is now executed only if `pdf.weight` is FALSE, as in the other
case `dens.max` is already computed elsewhere.