Clean up code to restore the graphical settings in plot_DoseResponseCurve().
This ensures that the par() settings modified inside the function (only cex
at the moment) are restored to their previous values in all cases, not just
if `plot_extended && !plot_singlePanels` like before.

The mfrow and mfcol settings need to be restored if layout() is called, as
that function modifies them, and not restoring them leaves the graphical
device in an unstable state that causes errors when running the examples in
the R CMD check phase.

This change affects the analyse_SAR.CWOSL() snapshot, as the cex setting is
now correctly reset to 2 after the dose-response curve plots. Therefore,
this also affects the plots produced by analyse_pIRIRSequence().

The graphical snapshots for plot_DoseResponseCurve() have changed a bit,
but there is no visual difference in the plots.