Add a workaround for R-devel failure.
Since today, R-devel has changed how some methods are searched if applied
to S4 objects. One of the affected functions is logLik(), which is called
from BIC(). When BIC() is called on an S4 object from the bbmle package,
it fails with this error:

Error in UseMethod("logLik") :
  no applicable method for 'logLik' applied to an object of class "mle2"
Calls: calc_MaxDose -> calc_MinDose -> BIC -> BIC.default -> ll

This is the relevant commit:
https://github.com/wch/r-source/commit/460a0f00cae8a3a850126a3c586c6d61fea60454

Importing explicitly the stats4::logLik() seems to be enough to work
around this issue.