Simplify code in analyse_Al2O3C_ITC() by using data.tables.
The change in the snapshot is due to the fact that we do not access the
RLum.Analysis object via [] anymore when computing net_SIGNAL. When [] is
used on an RLum.Analysis object is, a new UID is created for the resulting
object, and that alters the random seed. For example:

  waldo::compare(data_ITC[1], data_ITC[1])
  # `old@.uid`: "f3e64ef0cc41f127"
  # `new@.uid`: "83b299fae184de15"

Therefore, when we call fit_DoseResponseCurve() we have a different random
seed as before, so the resulting De.MC value is slightly different.