Clean up the use_DRAC() tests.
This adds a mock for readline() to prevent the test blocking for user input
when run interactively. Because readline() is a base function, the package
must provide a binding for it so the mock can be installed during tests
(when the namespace is locked and new bindings cannot be created).