Fix crash and hang in calc_AliquotSize().
In case the min and max grain sizes provided are the same, we skip the loop
to compute the MC sample mean, because the result in that case is already
known, and looping would cause the function to hang if grain.size contained
small values.

Moreover, grain.size is no longer allowed to be zero.