test_fit_DoseResponseCurve: Fit numerical results.
I have been seeing this discrepancy for a couple of months on my machine,
but from today it has started also on CI. I checked that the same error
is present when running 0.9.26, which means that the change is either in
some other package or on some system library. Why this error triggered at
different times is a mystery.