plot_RLum.Analysis: Restore functionality for combining plots.
This effectively reverts e4444847 without any adverse results in the set
of graphical snapshots currently available. Actually, this improves the
snapshots for analyse_FadingMeasurement(), which were affected by the same
issue but it went unnoticed when the snapshots were initially generated
a few days after that buggy commit (in 8052eb2e).